requirements.txt
# Core framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
websockets==14.1
pydantic==2.10.4
pydantic-settings==2.7.1

# Speech-to-text
deepgram-sdk==3.8.1
openai-whisper==20240930

# Text-to-speech
elevenlabs==1.17.0
pyttsx3==2.98

# LLM and orchestration
openai==1.58.1
langchain==0.3.14
langchain-openai==0.3.0
langchain-community==0.3.14
langgraph==0.2.62

# Vector store and embeddings
faiss-cpu==1.9.0.post1
chromadb==0.5.23
sentence-transformers==3.3.1

# Telephony
twilio==9.4.0

# Audio processing
numpy==2.2.1
soundfile==0.12.1
librosa==0.10.2.post1
webrtcvad==2.0.10

# Utilities
python-dotenv==1.0.1
httpx==0.28.1
aiofiles==24.1.0
python-multipart==0.0.20

# Logging and metrics
structlog==24.4.0
prometheus-client==0.21.1

# Testing
pytest==8.3.4
pytest-asyncio==0.25.0
pytest-cov==6.0.0
httpx==0.28.1

# Linting (dev)
ruff==0.8.6
